# try to make arithmetic work in datavalue2