--- a/common/datavalues2/src/macros.rs
+++ b/common/datavalues2/src/macros.rs
@@ -188,6 +188,67 @@ macro_rules! with_match_primitive_type {
 }
 
 #[macro_export]
+macro_rules! with_match_date_type_error {
+    (
+         $key_type:expr, | $_:tt $T:ident | $body:tt
+    ) => {{
+        macro_rules! __with_ty__ {
+            ( $_ $T:ident ) => {
+                $body
+            };
+        }
+
+        match $key_type {
+            TypeID::Date16 => __with_ty__! { u16},
+            TypeID::Date32 => __with_ty__! { i32},
+            TypeID::DateTime32 => __with_ty__! { u32},
+            TypeID::DateTime64 => __with_ty__! { u64},
+            v => Err(ErrorCode::BadDataValueType(format!(
+                "Ops is not support on datatype: {:?}",
+                v
+            ))),
+        }
+    }};
+}
+
+// doesn't include Bool and String
+#[macro_export]
+macro_rules! with_match_numeric_type_error {
+    (
+         $key_type:expr, | $_:tt $T:ident | $body:tt
+    ) => {{
+        macro_rules! __with_ty__ {
+            ( $_ $T:ident ) => {
+                $body
+            };
+        }
+
+        match $key_type {
+            TypeID::Int8 => __with_ty__! { i8 },
+            TypeID::Int16 => __with_ty__! { i16 },
+            TypeID::Int32 => __with_ty__! { i32 },
+            TypeID::Int64 => __with_ty__! { i64 },
+            TypeID::UInt8 => __with_ty__! { u8 },
+            TypeID::UInt16 => __with_ty__! { u16 },
+            TypeID::UInt32 => __with_ty__! { u32 },
+            TypeID::UInt64 => __with_ty__! { u64 },
+            TypeID::Float32 => __with_ty__! { f32 },
+            TypeID::Float64 => __with_ty__! { f64 },
+
+            TypeID::Date16 => __with_ty__! { u16},
+            TypeID::Date32 => __with_ty__! { i32},
+            TypeID::DateTime32 => __with_ty__! { u32},
+            TypeID::DateTime64 => __with_ty__! { u64},
+
+            v => Err(ErrorCode::BadDataValueType(format!(
+                "Ops is not support on datatype: {:?}",
+                v
+            ))),
+        }
+    }};
+}
+
+#[macro_export]
 macro_rules! match_data_type_apply_macro_ca {
     ($self:expr, $macro:ident, $macro_string:ident, $macro_bool:ident $(, $opt_args:expr)*) => {{