--- a/common/datavalues2/src/prelude.rs
+++ b/common/datavalues2/src/prelude.rs
@@ -17,8 +17,8 @@ use common_arrow::arrow::array::BinaryArray;
 use common_arrow::arrow::array::MutableBinaryArray;
 
 pub use crate::columns::*;
+pub use crate::convert::*;
 pub use crate::data_value::DFTryFrom;
-pub use crate::macros::*;
 pub use crate::scalars::*;
 pub use crate::types::*;
 pub use crate::utils::*;