--- a/common/functions/src/scalars/arithmetics/arithmetic.rs
+++ b/common/functions/src/scalars/arithmetics/arithmetic.rs
@@ -12,23 +12,32 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use common_datavalues::prelude::*;
-use common_exception::Result;
-
-use crate::scalars::function_factory::FunctionFactory;
+use crate::scalars::ArithmeticDivFunction;
+use crate::scalars::ArithmeticIntDivFunction;
+use crate::scalars::ArithmeticMinusFunction;
 use crate::scalars::ArithmeticModuloFunction;
-
-pub trait ArithmeticTrait {
-    fn arithmetic(columns: &DataColumnsWithField) -> Result<DataColumn>;
-}
+use crate::scalars::ArithmeticMulFunction;
+use crate::scalars::ArithmeticNegateFunction;
+use crate::scalars::ArithmeticPlusFunction;
+use crate::scalars::Function2Factory;
 
 #[derive(Clone)]
 pub struct ArithmeticFunction;
 
 impl ArithmeticFunction {
-    pub fn register(factory: &mut FunctionFactory) {
+    pub fn register(factory: &mut Function2Factory) {
+        factory.register_arithmetic("negate", ArithmeticNegateFunction::desc());
+        factory.register_arithmetic("+", ArithmeticPlusFunction::desc());
+        factory.register_arithmetic("plus", ArithmeticPlusFunction::desc());
+        factory.register_arithmetic("-", ArithmeticMinusFunction::desc());
+        factory.register_arithmetic("minus", ArithmeticMinusFunction::desc());
+        factory.register_arithmetic("*", ArithmeticMulFunction::desc());
+        factory.register_arithmetic("multiply", ArithmeticMulFunction::desc());
+        factory.register_arithmetic("/", ArithmeticDivFunction::desc());
+        factory.register_arithmetic("divide", ArithmeticDivFunction::desc());
+        factory.register_arithmetic("div", ArithmeticIntDivFunction::desc());
         factory.register_arithmetic("%", ArithmeticModuloFunction::desc());
         factory.register_arithmetic("modulo", ArithmeticModuloFunction::desc());
-        //factory.register_arithmetic("div", IntDivFunctionImpl::desc());
+        factory.register_arithmetic("mod", ArithmeticModuloFunction::desc());
     }
 }