--- a/common/functions/src/scalars/arithmetics/arithmetic_intdiv.rs
+++ b/common/functions/src/scalars/arithmetics/arithmetic_intdiv.rs
@@ -12,92 +12,39 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+use std::fmt;
 use std::marker::PhantomData;
 
-use common_datavalues::prelude::*;
-use common_datavalues::DataTypeAndNullable;
+use common_datavalues2::prelude::*;
+use common_datavalues2::with_match_primitive_type;
 use common_exception::ErrorCode;
 use common_exception::Result;
-use num::cast::AsPrimitive;
+use num_traits::AsPrimitive;
 
-use super::arithmetic::ArithmeticTrait;
-use crate::scalars::function_factory::ArithmeticDescription;
 use crate::scalars::function_factory::FunctionFeatures;
-use crate::scalars::BinaryArithmeticFunction;
-use crate::scalars::Function;
-use crate::scalars::Monotonicity;
-use crate::try_binary_arithmetic;
-use crate::with_match_primitive_type;
-
-#[derive(Clone)]
-pub struct ArithmeticIntDiv<T, D, R> {
-    t: PhantomData<T>,
-    d: PhantomData<D>,
-    r: PhantomData<R>,
-}
-
-impl<T, D, R> ArithmeticTrait for ArithmeticIntDiv<T, D, R>
-where
-    f64: AsPrimitive<R>,
-    T: DFPrimitiveType + AsPrimitive<f64>,
-    D: DFPrimitiveType + AsPrimitive<f64> + num::One,
-    R: DFPrimitiveType,
-    DFPrimitiveArray<R>: IntoSeries,
-    R: Into<DataValue>,
-{
-    fn arithmetic(columns: &DataColumnsWithField) -> Result<DataColumn> {
-        try_binary_arithmetic! {
-            columns[0].column(),
-            columns[1].column(),
-            f64,
-            |l: f64, r:f64| {
-                if std::intrinsics::unlikely(r == 0.0) {
-                    return Err(ErrorCode::BadArguments("Division by zero"));
-                }
-                Ok(AsPrimitive::<R>::as_(l / r))
-            },
-            |lhs: &DFPrimitiveArray<T>, r: f64| {
-                if r == 0.0 {
-                    return Err(ErrorCode::BadArguments("Division by zero"));
-                }
-                Ok(unary(lhs, |l| {
-                    AsPrimitive::<R>::as_(AsPrimitive::<f64>::as_(l) / r)
-                })
-                .into())
-            }
-        }
-    }
-}
+use crate::scalars::Arithmetic2Description;
+use crate::scalars::Function2;
 
 pub struct ArithmeticIntDivFunction;
 
 impl ArithmeticIntDivFunction {
     pub fn try_create_func(
         _display_name: &str,
-        args: &[DataTypeAndNullable],
-    ) -> Result<Box<dyn Function>> {
-        let left_type = &args[0].data_type();
-        let right_type = &args[1].data_type();
-        let op = DataValueBinaryOperator::IntDiv;
-        let error_fn = || -> Result<Box<dyn Function>> {
-            Err(ErrorCode::BadDataValueType(format!(
-                "DataValue Error: Unsupported arithmetic ({:?}) {} ({:?})",
-                left_type, op, right_type
-            )))
-        };
+        args: &[&DataTypePtr],
+    ) -> Result<Box<dyn Function2>> {
+        let left_type = remove_nullable(args[0]).data_type_id();
+        let right_type = remove_nullable(args[1]).data_type_id();
 
-        // error on any non-numeric type
-        if !left_type.is_numeric() || !right_type.is_numeric() {
-            return error_fn();
+        let error_fn = || -> Result<Box<dyn Function2>> {
+            Err(ErrorCode::BadDataValueType(format!(
+                "DataValue Error: Unsupported arithmetic ({:?}) div ({:?})",
+                left_type, right_type
+            )))
         };
 
         with_match_primitive_type!(left_type, |$T| {
             with_match_primitive_type!(right_type, |$D| {
-                let result_type = <($T, $D) as ResultTypeOfBinary>::IntDiv::data_type();
-                BinaryArithmeticFunction::<ArithmeticIntDiv::<$T,$D, <($T, $D) as ResultTypeOfBinary>::IntDiv>>::try_create_func(
-                    op,
-                    result_type,
-                )
+                Ok(Box::new(IntDivFunctionImpl::<$T, $D, <($T, $D) as ResultTypeOfBinary>::IntDiv>::default()))
             }, {
                 error_fn()
             })
@@ -106,16 +53,57 @@ impl ArithmeticIntDivFunction {
         })
     }
 
-    pub fn desc() -> ArithmeticDescription {
-        ArithmeticDescription::creator(Box::new(Self::try_create_func)).features(
+    pub fn desc() -> Arithmetic2Description {
+        Arithmetic2Description::creator(Box::new(Self::try_create_func)).features(
             FunctionFeatures::default()
                 .deterministic()
                 .monotonicity()
                 .num_arguments(2),
         )
     }
+}
 
-    pub fn get_monotonicity(_args: &[Monotonicity]) -> Result<Monotonicity> {
-        Ok(Monotonicity::default())
+#[derive(Clone, Default)]
+pub struct IntDivFunctionImpl<L, R, O> {
+    l: PhantomData<L>,
+    r: PhantomData<R>,
+    o: PhantomData<O>,
+}
+
+impl<L, R, O> Function2 for IntDivFunctionImpl<L, R, O>
+where
+    f64: AsPrimitive<O>,
+    L: PrimitiveType + AsPrimitive<f64>,
+    R: PrimitiveType + AsPrimitive<f64>,
+    O: IntegerType + ToDataType,
+{
+    fn name(&self) -> &str {
+        "IntDivFunctionImpl"
+    }
+
+    fn return_type(&self, _args: &[&DataTypePtr]) -> Result<DataTypePtr> {
+        Ok(O::to_data_type())
+    }
+
+    fn eval(&self, columns: &ColumnsWithField, _input_rows: usize) -> Result<ColumnRef> {
+        let left = ColumnViewerIter::<L>::try_create(columns[0].column())?;
+        let right = ColumnViewerIter::<R>::try_create(columns[1].column())?;
+        let mut col_builder = MutablePrimitiveColumn::<O>::with_capacity(left.size);
+        for (l, r) in left.zip(right) {
+            let l = l.to_owned_scalar().as_();
+            let r = r.to_owned_scalar().as_();
+            if std::intrinsics::unlikely(r == 0.0) {
+                return Err(ErrorCode::BadArguments("Division by zero"));
+            }
+            let o = AsPrimitive::<O>::as_(l / r);
+            col_builder.append_value(o);
+        }
+        Ok(col_builder.to_column())
+    }
+}
+
+impl<L, R, O> fmt::Display for IntDivFunctionImpl<L, R, O> {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "div")
     }
 }