… common/codegen/src/writes/arithmetics_type.rs
… common/datavalues2/src/types/arithmetics_type.rs
… common/functions/src/scalars/arithmetics/arithmetic_modulo.rs
… common/functions/src/scalars/arithmetics/arithmetic_mul.rs
… common/functions/src/scalars/arithmetics/arithmetic_negate.rs
… common/functions/src/scalars/arithmetics/arithmetic_plus.rs
… common/functions/src/scalars/arithmetics/binary_arithmetic.rs
… common/functions/src/scalars/arithmetics/binary_arithmetic2.rs
… common/functions/src/scalars/arithmetics/interval.rs
… common/functions/src/scalars/arithmetics/macros.rs
… common/functions/src/scalars/arithmetics/mod.rs
… common/functions/src/scalars/arithmetics/unary_arithmetic.rs
… common/functions/src/scalars/arithmetics/utils.rs
… common/functions/src/scalars/dates/interval_function.rs
… common/functions/src/scalars/dates/mod.rs
… common/functions/src/scalars/expressions/binary.rs
… common/functions/src/scalars/expressions/cast_with_type.rs
… common/functions/src/scalars/expressions/mod.rs
… common/functions/src/scalars/expressions/unary.rs
… common/functions/src/scalars/function2_adapter.rs
… common/functions/src/scalars/function2_factory.rs
… common/functions/src/scalars/function_factory.rs
… common/functions/src/scalars/maths/math.rs
… common/functions/tests/it/scalars/arithmetics.rs
… common/planners/tests/it/plan_builder.rs
… tests/suites/0_stateless/02_0012_function_datetimes.result
… tests/suites/0_stateless/02_0012_function_datetimes.sql