@zhyass 2022-01-26
CC @sundy-li , `cast` performance drop 10%～15%, `plus` performance drop 80%

Performance:

Before:
```
mysql> SELECT count(toUInt32(number)) FROM numbers(1000000000);
+-------------------------+
| count(toUInt32(number)) |
+-------------------------+
|              1000000000 |
+-------------------------+
1 row in set (0.19 sec)
Read 1000000000 rows, 8 GB in 0.186 sec., 5.39 billion rows/sec., 43.1 GB/sec.

mysql> SELECT count(toUInt32(number) + toInt64(number)) FROM numbers(1000000000);
+---------------------------------------------+
| count((toUInt32(number) + toInt64(number))) |
+---------------------------------------------+
|                                  1000000000 |
+---------------------------------------------+
1 row in set (0.43 sec)
Read 1000000000 rows, 8 GB in 0.420 sec., 2.38 billion rows/sec., 19.07 GB/sec.
```

After:
```
mysql> SELECT count(toUInt32(number)) FROM numbers(1000000000);
+-------------------------+
| count(toUInt32(number)) |
+-------------------------+
|              1000000000 |
+-------------------------+
1 row in set (0.22 sec)
Read 1000000000 rows, 8 GB in 0.216 sec., 4.64 billion rows/sec., 37.1 GB/sec.

mysql> SELECT count(toUInt32(number) + toInt64(number)) FROM numbers(1000000000);
+---------------------------------------------+
| count((toUInt32(number) + toInt64(number))) |
+---------------------------------------------+
|                                  1000000000 |
+---------------------------------------------+
1 row in set (2.21 sec)
Read 1000000000 rows, 8 GB in 2.203 sec., 454.02 million rows/sec., 3.63 GB/sec.
```