@zhyass 2022-01-27
TODO: interval arithmetic will be pushed in next pr.